runner_ui/requirements.txt
pywebview==5.4
requests==2.32.4
protobuf==6.31.1
construct==2.10.70
pytest==8.4.1
pytest-html==4.1.1
allure-pytest==2.14.3
pyinstaller==6.14.2

